requirements.txt
pandas
psycopg2
streamlit
requests
apache-airflow
